Add a UDP client to allow for comparison
It has the same command-line interface than tcpclient, and the same
Poisson behaviour.  The notion of "connection" does not make much sense
with UDP: here, sending 2 queries on a single connection means that they
will use the same UDP source port.

The implementation is a bit simpler than with TCP, because we always get a
single DNS message when reading on a socket.

Note that we don't use bufferevents here, because they are not really
adapted to UDP (the semantic of bufferevents is stream-oriented in
nature).